+++ MekHQ/resources/mekhq/resources/GUI.properties
@@ -97,7 +97,7 @@ changeMDRank.text=Change Manei Domini Rank
 changePrimaryDesignation.text=Change Primary Designation
 changeSecondaryDesignation.text=Change Secondary Designation
 changeStatus.text=Change Status
-imprison.text=Imprison
+makePrisonerOfWar.text=Make Prisoner of War
 free.text=Free
 execute.text=Execute
 jettison.text=Jettison
