+++ MekHQ/src/mekhq/campaign/personnel/marriage/AbstractMarriage.java
@@ -30,7 +30,7 @@ import mekhq.campaign.log.PersonalLogger;
 import mekhq.campaign.personnel.Person;
 import mekhq.campaign.personnel.enums.MergingSurnameStyle;
 import mekhq.campaign.personnel.enums.RandomMarriageMethod;
-import mekhq.campaign.personnel.randomEvents.prisoners.enums.PrisonerStatus;
+import mekhq.campaign.randomEvents.prisoners.enums.PrisonerStatus;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
