+++ MekHQ/src/mekhq/campaign/randomEvents/RandomEventLibraries.java
@@ -1,9 +1,9 @@
-package mekhq.campaign.personnel.randomEvents;
+package mekhq.campaign.randomEvents;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import mekhq.campaign.personnel.randomEvents.prisoners.records.PrisonerEventData;
-import mekhq.campaign.personnel.randomEvents.prisoners.yaml.PrisonerEventDataWrapper;
+import mekhq.campaign.randomEvents.prisoners.records.PrisonerEventData;
+import mekhq.campaign.randomEvents.prisoners.yaml.PrisonerEventDataWrapper;
 
 import java.io.File;
 import java.io.IOException;
