+++ MekHQ/src/mekhq/campaign/randomEvents/prisoners/PrisonerRansomEvent.java
@@ -1,4 +1,4 @@
-package mekhq.campaign.personnel.randomEvents.prisoners;
+package mekhq.campaign.randomEvents.prisoners;
 
 import mekhq.campaign.Campaign;
 import mekhq.campaign.finances.Money;
