+++ MekHQ/src/mekhq/campaign/randomEvents/prisoners/enums/ResponseQuality.java
@@ -1,4 +1,4 @@
-package mekhq.campaign.personnel.randomEvents.prisoners.enums;
+package mekhq.campaign.randomEvents.prisoners.enums;
 
 public enum ResponseQuality {
     RESPONSE_NEUTRAL, RESPONSE_POSITIVE, RESPONSE_NEGATIVE
