+++ MekHQ/unittests/mekhq/campaign/personnel/enums/PrisonerCaptureStyleTest.java
@@ -19,7 +19,7 @@
 package mekhq.campaign.personnel.enums;
 
 import mekhq.MekHQ;
-import mekhq.campaign.personnel.randomEvents.prisoners.enums.PrisonerCaptureStyle;
+import mekhq.campaign.randomEvents.prisoners.enums.PrisonerCaptureStyle;
 import org.junit.jupiter.api.Test;
 
 import java.util.ResourceBundle;
